Split model.rs into separate files
Now the structure follows the "attribute" module. The implementations
of the "attribute" and "model" are linked. The modules are
easier to understand if their implementation structures are similar.
